refactor: `auth` 패키지로 토큰 인증 관련 클래스 이동
