+++ src/main/java/roomescape/application/member/MemberService.java
@@ -7,6 +7,7 @@ import roomescape.application.member.dto.request.MemberLoginRequest;
 import roomescape.application.member.dto.request.MemberRegisterRequest;
 import roomescape.application.member.dto.response.MemberResponse;
 import roomescape.application.member.dto.response.TokenResponse;
+import roomescape.auth.TokenManager;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberRepository;
 import roomescape.domain.role.MemberRole;
