+++ src/main/java/roomescape/auth/LoginMemberId.java
@@ -1,4 +1,4 @@
-package roomescape.application.member;
+package roomescape.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
