+++ src/main/java/roomescape/auth/TokenManager.java
@@ -1,4 +1,4 @@
-package roomescape.application.member;
+package roomescape.auth;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
