+++ src/main/java/roomescape/presentation/reservation/ReservationController.java
@@ -10,11 +10,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.application.member.LoginMemberId;
 import roomescape.application.reservation.ReservationService;
 import roomescape.application.reservation.dto.request.AdminReservationRequest;
 import roomescape.application.reservation.dto.request.ReservationRequest;
 import roomescape.application.reservation.dto.response.ReservationResponse;
+import roomescape.auth.LoginMemberId;
 
 @RestController
 public class ReservationController {
