+++ src/test/java/roomescape/auth/TokenManagerTest.java
@@ -1,4 +1,4 @@
-package roomescape.application.member;
+package roomescape.auth;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
