+++ src/test/java/roomescape/presentation/reservation/ReservationTimeControllerTest.java
@@ -19,10 +19,10 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
-import roomescape.application.member.TokenManager;
 import roomescape.application.reservation.ReservationTimeService;
 import roomescape.application.reservation.dto.request.ReservationTimeRequest;
 import roomescape.application.reservation.dto.response.ReservationTimeResponse;
+import roomescape.auth.TokenManager;
 import roomescape.presentation.ControllerTest;
 
 @WebMvcTest(ReservationTimeController.class)
